Testing for transaction gossip and receive + block creation
