+++ blockchain.py
@@ -1,8 +1,9 @@
 import binascii
 import random
-from collections import defaultdict
-import pickle
+import sys
 import logging
+from collections import defaultdict
+from typing import Dict
 
 from ipv8.community import Community, CommunitySettings
 from ipv8.lazy_community import lazy_wrapper
@@ -10,26 +11,19 @@ from ipv8.messaging.payload_dataclass import dataclass
 from ipv8.types import Peer
 
 from block import Block
+from transaction import Transaction
 
 # Amount of peers to send message to
 k = 2
 
-logging.basicConfig(filename='blockchain.log', level=logging.INFO, format='%(asctime)s %(levelname)s: %(message)s')
-
-
-@dataclass(msg_id=1)  # The value 1 identifies this message and must be unique per community
-class Transaction:
-    sender: bytes
-    receiver: bytes
-    amount: int
-    nonce: int = 1
-    ttl: int = 3
+logging.basicConfig(stream=sys.stdout, level=logging.INFO, format='%(asctime)s %(levelname)s: %(message)s')
 
 
 @dataclass(msg_id=2)
 class BlockMessage:
     hash: str
-    block: bytes
+    block: Block
+    ttl: int = 3
 
 
 class MyCommunity(Community):
@@ -41,50 +35,51 @@ class MyCommunity(Community):
         self.max_messages = 5
         self.executed_checks = 0
 
-        self.pending_txs = []
-        self.finalized_txs = []
+        self.pending_txs: Dict[bytes, Transaction] = {}
+        self.finalized_txs: Dict[bytes, Transaction] = {}
         self.balances = defaultdict(lambda: 1000)
         self.blocks = []  # List to store finalized blocks
         self.current_block = Block('0')  # Current working block
 
         self.add_message_handler(Transaction, self.on_transaction)
-        self.add_message_handler(BlockMessage, self.on_block)
+        self.add_message_handler(BlockMessage, self.receive_block)
 
     def started(self, id) -> None:
         logging.info('Community started')
+
         # Testing purpose
         if id == 1:
-            self.register_task("tx_create", self.create_transaction, delay=1, interval=5)
-        # WIP: Check if this is the right way to check transactions
-        self.register_task("check_txs", self.check_transactions, delay=1, interval=5)
+            random_transaction_interval = random.randint(5, 10)
+            self.register_task("tx_create", self.create_transaction, delay=1, interval=random_transaction_interval)
+
+        random_check_interval = random.randint(5, 10)
+        self.register_task("check_txs", self.block_creation, delay=1, interval=random_check_interval)
 
     def peers_found(self):
         return len(self.get_peers()) > 0
 
-    def get_peer_id(self, peer: Peer = None):
+    def get_peer_id(self, peer: Peer = None) -> str:
         return binascii.hexlify(peer.mid).decode()
 
-    def create_transaction(self):
+    def create_transaction(self) -> None:
         if not self.peers_found():
-            print(f'[Node {self.get_peer_id(self.my_peer)}] No peers found')
             logging.info(f'[Node {self.get_peer_id(self.my_peer)}] No peers found')
             return
 
-        print(f'[Node {self.get_peer_id(self.my_peer)}] Creating transaction')
         logging.info(f'[Node {self.get_peer_id(self.my_peer)}] Creating transaction')
-        peer = random.choice([i for i in self.get_peers()])
-        peer_id = peer.mid
+        receiver_peer = random.choice([i for i in self.get_peers()])
 
-        tx = Transaction(self.my_peer.mid,
-                         peer_id,
-                         10,
-                         self.counter)
-        self.counter += 1
-        print(f'[Node {self.get_peer_id(self.my_peer)}] Sending transaction {tx.nonce} to {self.get_peer_id(peer)}')
+        # ttl = 3 when creating a tx
+        tx = Transaction(self.my_peer.mid, receiver_peer.mid, 10, nonce=self.counter)
+        tx.public_key = self.crypto.key_to_bin(self.my_peer.key.pub())
+        # should we sigh the hash of tx or the whole tx?
+        tx.signature = self.crypto.create_signature(self.my_peer.key, tx.get_tx_bytes())
         logging.info(
-            f'[Node {self.get_peer_id(self.my_peer)}] Sending transaction {tx.nonce} to {self.get_peer_id(peer)}')
-
-        self.ez_send(peer, tx)
+            f'[Node {self.get_peer_id(self.my_peer)}] Sending transaction {tx.nonce} to {self.get_peer_id(receiver_peer)}')
+        self.pending_txs[tx.get_tx_hash()] = tx
+        # and this is correct that initially we send this tx only to receiver?
+        self.ez_send(receiver_peer, tx)
+        self.counter += 1
 
         # WIP: We need this?
         # if self.counter > self.max_messages:
@@ -92,72 +87,85 @@ class MyCommunity(Community):
         #     self.stop()
         #     return
 
-    def check_transactions(self):
-        print(f'[Node {self.get_peer_id(self.my_peer)}] Checking transactions')
-        logging.info(f'[Node {self.get_peer_id(self.my_peer)}] Checking transactions')
+    def block_creation(self):
+        # WIP: get every known peer, not only the ones im connected to
+        peers = self.get_peers()
+        peers.append(self.my_peer)
+        selectedPeer = random.choice(peers)
 
-        for tx in self.pending_txs:
-            if self.balances[tx.sender] - tx.amount >= 0:
-                self.balances[tx.sender] -= tx.amount
-                self.balances[tx.receiver] += tx.amount
-                self.pending_txs.remove(tx)
-                self.finalized_txs.append(tx)
-                self.current_block.add_transaction(tx)
+        print(selectedPeer.mid, self.my_peer.mid)
+        if not selectedPeer.mid == self.my_peer.mid:
+            return
 
-                if self.current_block.is_full():
-                    print('Block is full')
-                    self.finalize_and_broadcast_block()
-                    break
+        logging.info(f'[Node {self.get_peer_id(self.my_peer)}] is creating a block')
+        for tx_hash in list(self.pending_txs.keys()):
+            tx = self.pending_txs.pop(tx_hash)
+            self.finalized_txs[tx_hash] = tx
+            self.current_block.add_transaction(tx)
+
+            if self.current_block.is_full():
+                logging.info(f'[Node {self.get_peer_id(self.my_peer)}] Chosen one')
+                self.finalize_and_broadcast_block()
+                break
 
     @lazy_wrapper(Transaction)
-    async def on_transaction(self, peer: Peer, payload: Transaction) -> None:
+    async def on_transaction(self, peer: Peer, tx: Transaction) -> None:
         my_id = self.get_peer_id(self.my_peer)
+        logging.info(f'[Node {my_id}] received transaction {tx.nonce} from {self.get_peer_id(peer)}')
 
-        print(f'[Node {my_id}] Received transaction', payload.nonce, 'from', self.get_peer_id(peer))
-        logging.info(f'[Node {my_id}] Received transaction {payload.nonce} from {self.get_peer_id(peer)}')
+        tx_hash = tx.get_tx_hash()
+        # if we already have this tx we do nothing
+        # here if we have the same txs received from different peers - we choose the one that was sent earlier?
+        if tx_hash in self.finalized_txs or tx_hash in self.pending_txs:
+            return
 
-        # Add to pending transactions
-        if (payload.sender, payload.nonce) not in [(tx.sender, tx.nonce) for tx in self.finalized_txs] and (
-                payload.sender, payload.nonce) not in [(tx.sender, tx.nonce) for tx in self.pending_txs]:
-            self.pending_txs.append(payload)
+        # if the signature of tx is not valid we do nothing
+        # todo need to get PublicKey object from bytes
+        if not self.crypto.is_valid_signature(self.crypto.key_from_public_bin(tx.public_key), tx.get_tx_bytes(), tx.signature):
+            logging.info(f'[Node {my_id}]: signature incorrect')
+            return
 
-        # If we are connected to more than k peers, we can gossip
-        # only if the ttl is greater than 0
-        if len(self.get_peers()) > k and payload.ttl > 0:
-            payload.ttl -= 1
-
+        logging.info(f'[Node {my_id}]: signature correct')
+        self.pending_txs[tx.get_tx_hash()] = tx
+        if tx.ttl > 0:
+            tx.ttl -= 1
             # push gossip to k random peers
-            get_peers_to_distribute = random.sample(self.get_peers(), k)
+            get_peers_to_distribute = random.sample(self.get_peers(), min(k, len(self.get_peers())))
             for peer in get_peers_to_distribute:
-                self.ez_send(peer, payload)
+                self.ez_send(peer, tx)
 
     @lazy_wrapper(BlockMessage)
-    async def on_block(self, peer: Peer, payload: BlockMessage) -> None:
-        unserialized_block = pickle.loads(payload.block)
-        print(unserialized_block)
+    async def receive_block(self, peer: Peer, payload: BlockMessage) -> None:
+        print('----------on block----------')
+        print(payload)
+
+        print(payload.block.previous_hash)
+        print(payload.block.merkle_hash)
+        print(payload.block.transactions)
 
         # Check if the block is already in the chain
-        if payload.hash not in [block.merkle_tree.get_root_hash() for block in self.blocks]:
-            print(f'Block {payload.hash} not in my chain {self.get_peer_id(peer)}')
-            logging.info(f'Block {payload.hash} not in my chain {self.get_peer_id(peer)}')
-            self.blocks.append(unserialized_block)
+        # if payload.hash not in [block.get_merkle_hash() for block in self.blocks]:
+        #     print(f'Block {payload.hash} not in my chain {self.get_peer_id(peer)}')
+        #     logging.info(f'Block {payload.hash} not in my chain {self.get_peer_id(peer)}')
+        #     self.blocks.append(payload.block)
 
-        # WIP: Necessary check?
-        # Check if the previous block hash is indeed the past block
+        # Remove block transactions from my pending_txs list
+        # for tx in unserialized_block.transactions:
+        #     if (tx.sender, tx.nonce) in [(tx.sender, tx.nonce) for tx in self.pending_txs]:
+        #         self.pending_txs.remove(tx)
 
     def finalize_and_broadcast_block(self):
-        self.current_block.merkle_tree.recalculate_tree()
-        new_block_hash = self.current_block.merkle_tree.get_root_hash()
+        self.current_block.update_tree()
+        new_block_hash = self.current_block.get_merkle_hash()
         print(f'New block hash: {new_block_hash}')
         logging.info(f'New block hash: {new_block_hash}')
         self.blocks.append(self.current_block)
-        serialized_block = pickle.dumps(self.current_block)
+
+        self.broadcast_block(new_block_hash, self.current_block)
         self.current_block = Block(new_block_hash)
 
-        self.broadcast_new_block(new_block_hash, serialized_block)
-
-    def broadcast_new_block(self, block_hash, serialized_block):
-        blockMessage = BlockMessage(block_hash, serialized_block)
+    def broadcast_block(self, block_hash: str, block: Block):
+        blockMessage = BlockMessage(block_hash, block)
 
         for peer in self.get_peers():
             self.ez_send(peer, blockMessage)
